First steps towards non-blocking mode.

+++ parcalc/parcalc.py
@@ -55,6 +55,7 @@ import os
 import tempfile
 import shutil
 import copy
+from subprocess import check_output
 
 class ClusterVasp(Vasp):
     '''
@@ -68,6 +69,11 @@ class ClusterVasp(Vasp):
         Vasp.__init__(self, **kwargs)
         self.nodes=nodes
         self.ppn=ppn
+        self.calc_running=False
+        if 'block' in kwargs :
+            self.block=block
+        else :
+            self.block=True
         
     def prepare_calc_dir(self):
         '''
@@ -76,13 +82,59 @@ class ClusterVasp(Vasp):
         The following code reflects just my particular setup.
         '''
         f=open("vasprun.conf","w")
-        f.write('NODES="nodes=%s:ppn=%d"' % (self.nodes, self.ppn))
+        f.write('NODES="nodes=%s:ppn=%d"\n' % (self.nodes, self.ppn))
+        f.write('BLOCK=%d\n' % (self.block,))
         #print(self.nodes, self.ppn)
         f.close()
+
+    def __calc_finished(self):
+        '''
+        Check if the lockfile is in the calculation directory.
+        It is removed by the script at the end regardless of the 
+        success of the calculation. This is totally tied to
+        implementation and you need to implement your own scheme!
+        '''
+        if not self.calc_running : return True
+        else:
+            # The calc is marked as running check if this is still true
+            # we use our unique job name to ask the queuing system
+            # about the status. We do it by external scripts.
+            # You need to write these scripts for your system.
+            
+   
+    def update(self, atoms):
+        if self.calc_running :
+            # we have started the calculation and have 
+            # nothing to read really. But we need to check
+            # first if this is still true.
+            
+
+        # We are not in the middle of calculation.
+        # Update as normal
+        self.update(atoms)
+        
    
     def calculate(self, atoms):
         self.prepare_calc_dir()
-        Vasp.calculate(self, atoms)
+        if self.block:
+            # just run, wait for the job and read in the results.
+            # Easy ...
+            Vasp.calculate(self, atoms)
+        else :
+            # This is tricky. We need to start the job and guard
+            # against it reading back possible old data from the
+            # directory - the queuing system may not even started
+            # the job when we get control back from the starting 
+            # script. Thus anything we read after invocation is 
+            # potentially garbage - even if it is converged 
+            # calculation data.
+            
+            # start the calculation expecting data reading functions 
+            # to fail at the end. We need to handle any exceptions
+            # and mark the calculation as not ready.
+            self.calc_running=True
+            try :
+                Vasp.calculate(self, atoms)
 
 
 
@@ -117,14 +169,33 @@ class __PCalcProc(Process):
     '''
     Internal helper class representing the calculation process isolated
     from the rest of the ASE script. The process (not thread) runs in 
-    the separate, temporary directory, created on-the-fly and removed at
-    the end. It is vital for the calculator to read in all the results 
-    after the run since the files will be removed as soon as the 
+    the separate directory, created on-the-fly and removed at the end 
+    if the cleanup is true and we are in blocking (default) mode.
+    In this mode it is vital for the calculator to read in all the 
+    results after the run since the files will be removed as soon as the 
     "calculate" function terminates. You can pass False to the cleanup
     argument to prevent the clean-up. This is very usefull for debuging.
+    In the non-blocking mode nothing is removed - since the calculation 
+    is presumably not finished.
+    
+    The procedure has two modes of operation: blocking and non-blocking.
+    
+    block=True  [default] the process waits without any time-out for the 
+                calculation to finish. This is great for short and simple 
+                calculations and quick testing. You run the job and get 
+                back your results.
+                
+    block=False the process starts the job and returns immediately. 
+                You need to return to the calculation and check if it
+                is finished and read in the results. Not every calculator 
+                supports this feature. At least VASP does. In this mode
+                the cleanup parameter is obviously ignored. You need to 
+                clean up yourself. This mode is great for the long, problematic,
+                production runs when you need to tweak the calculation
+                to make it converge etc.
     '''
     
-    def __init__(self, iq, oq, calc, prefix, cleanup=True):
+    def __init__(self, iq, oq, calc, prefix, cleanup=True, block=True):
         Process.__init__(self)
         self.calc=calc
         self.basedir=os.getcwd()
@@ -132,13 +203,14 @@ class __PCalcProc(Process):
         self.iq=iq
         self.oq=oq
         self.CleanUp=cleanup
+        self.block=block
     
     def run(self):
         wd=os.getcwd()
         os.chdir(self.place)
         n,system=self.iq.get()
         system.set_calculator(copy.deepcopy(self.calc))
-        
+        system.calc.working_dir=self.place
         #print("Start at :", self.place)
         if hasattr(self.calc, 'name') and self.calc.name=='Siesta':
             system.get_potential_energy()
@@ -147,13 +219,13 @@ class __PCalcProc(Process):
         
         #print("Finito: ", os.getcwd(), system.get_volume(), system.get_pressure())
         self.oq.put([n,system])
-        if self.CleanUp :
+        if self.block and self.CleanUp :
             system.calc.clean()
             os.chdir(wd)
             shutil.rmtree(self.place, ignore_errors=True)
 
 
-def ParCalculate(systems,calc,cleanup=True,prefix="Calc_"):
+def ParCalculate(systems,calc,cleanup=True,block=True,prefix="Calc_"):
     '''
     Run calculators in parallel for all systems. 
     Calculators are executed in isolated processes and directories.
@@ -170,7 +242,7 @@ def ParCalculate(systems,calc,cleanup=True,prefix="Calc_"):
         
     # Create workers    
     for s in sys:
-        __PCalcProc(iq, oq, calc, prefix=prefix, cleanup=cleanup).start()
+        __PCalcProc(iq, oq, calc, prefix=prefix, cleanup=cleanup, block=block).start()
 
     # Put jobs into the queue
     for n,s in enumerate(sys):
@@ -189,6 +261,8 @@ def ParCalculate(systems,calc,cleanup=True,prefix="Calc_"):
         res.append([n,s])
         #print("Got from oq:", n, s.get_volume(), s.get_pressure())
     return [r for ns,s in enumerate(sys) for nr,r in res if nr==ns]
+
+
 
 # Testing routines using VASP as a calculator in the cluster environment.
 # TODO: Make it calculator/environment agnostic
